+++ src/SxDate.h
@@ -29,7 +29,7 @@ public:
     QString get_to_string(QString format);
     /// \brief Get the the date year
     /// \return Year value. Ex 2021
-    qint8 get_year();
+    qint16 get_year();
     /// \brief Get the the date month
     /// \return Month value. Ex 9
     qint8 get_month();
@@ -41,7 +41,7 @@ public:
     // setters
     /// \brief Set the date year
     /// \param[in] value Year of the date
-    void set_year(qint8 value);
+    void set_year(qint16 value);
     /// \brief Set the date month
     /// \param[in] value Month of the date
     void set_month(qint8 value);
@@ -56,10 +56,10 @@ private:
     /// \brief Convert a number to string with a fixed number of digit
     /// \param[in] number Number to convert
     /// \param[in] digit Number of digits
-    QString number_to_str(qint8 number, qint8 digits=2);
+    QString number_to_str(qint16 number, qint8 digits=2);
 
 private:
-    qint8 m_year; ///< date year
+    qint16 m_year; ///< date year
     qint8 m_month; ///< date month
     qint8 m_day; ///< date day
